tidy(call-diagnostics): extract the pure analysis engine
Moves timeline normalization, issue detection and repair recommendation out of
diagnose.py into a sibling module, analysis.py. diagnose.py keeps workspace and
metrics-path resolution, CLI parsing, SQLite/JSONL loading, terminal printing
and HTML tracker rendering.

  skills/call-diagnostics/scripts/diagnose.py   968 -> 547
  skills/call-diagnostics/scripts/analysis.py     0 -> 459  (new)

Stays inside skills/call-diagnostics/ — this is feature-specific policy and is
not promoted into src/.

The move is verbatim; the engine is import-safe (no workspace resolution, no
sys.argv, no SQLite, no file IO, no printing, no HTML), and diagnose.py imports
the names so existing call sites are untouched.

Behavior-preserving, measured rather than asserted. A characterization harness
exercises merge_timeline / parse_ts / diagnose / categorize_issue /
analyze_patterns_and_repair over fixture calls and dumps a full snapshot:

  parent (main)                0b754a4f5fe46174
  HEAD via diagnose.py         0b754a4f5fe46174
  HEAD via analysis.py direct  0b754a4f5fe46174

Both entry points agree with the parent.

tests/call-diagnostics-analysis.test.py — 36 checks. Timeline merge/order/shape
including that duplicates are NOT de-duplicated; parse_ts on valid, offset,
invalid and empty input; the sub-10ms "returned too fast" rule and its
work/hang_up exemption; hallucination detection; categorization totality and
determinism; repair determinism and the empty-calls case; and a purity contract
that scans the engine source (docstring stripped) for sqlite3 / sys.argv /
resolve_workspace / HTML, asserts a silent import, and asserts diagnose.py does
not redefine any moved symbol.

PRE-EXISTING DEFECT FOUND, RECORDED, NOT FIXED (separate PR):
analyze_patterns_and_repair raises IndexError when a call carries `events: []`,
because `call.get("events", [{}])[0]` defaults only on a MISSING key, not an
empty list. Present on main; the harness and the suite both characterize the
crash so this extraction cannot silently change it.

Two of my own test assumptions were wrong and the code was right: the fast-fail
rule triggers on a sub-10ms return (not a slow one), and hallucination lines are
prefixed "sutando:" (not "agent:"). Fixed the tests, not the behavior.
